feat: Break the kindelia_core package in three
Separates the kindelia_lang into a different crate with  ast and parser.
Separates some components into a kindelia_common crate.
Some refactoring to decouple things and improve readability.
