+++ kindelia_client/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+kindelia_lang = { path="../kindelia_lang" }
 kindelia_core = { path="../kindelia_core" }
 
 reqwest = { version = "0.11", features = ["json"] }
