+++ kindelia_common/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "kindelia_common"
+version = "0.1.5"
+edition = "2021"
+description = "An efficient, secure cryptocomputer."
+license = "MIT"
+repository = "https://github.com/Kindelia/Kindelia"
+
+
+[dependencies]
+# nohash-hasher = "0.2.0" # inlined, because we need u128
+serde = { version = "1.0.137", features = ["derive"] }
+
+bit-vec = "0.6.3"
+primitive-types = { version = "0.11.1" }
+hex = "0.4"
+
+# Crypto
+secp256k1 = { version = "0.22.1", features = [
+  "rand-std",
+  "recovery",
+  "global-context",
+] }
+tiny-keccak = { version = "2.0.2", features = ["keccak"] }
+sha3 = "0.9.1"
