+++ kindelia_common/src/nohash_hasher.rs
