+++ kindelia_core/src/lib.rs
@@ -1,19 +1,15 @@
 #![allow(clippy::single_component_path_imports)]
 
 #[allow(non_snake_case)]
-pub mod NoHashHasher;
 pub mod api;
 pub mod bits;
-pub mod common;
 pub mod constants;
-pub mod crypto;
 pub mod hvm;
 pub mod net;
 pub mod node;
 pub mod util;
 pub mod config;
 pub mod persistence;
-pub mod parser;
 
 #[cfg(feature = "events")]
 pub mod events;
