+++ kindelia_lang/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod ast;
+pub mod parser;
