+++ kindelia_server/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+kindelia_common = { path = "../kindelia_common" }
+kindelia_lang = { path = "../kindelia_lang" }
 kindelia_core = { path = "../kindelia_core" }
 
 bit-vec = "0.6.3"
